TranscriptUtilities load strongly typed activities
Needed to update `TranscriptUtilities` which has/needs its own
deserialization logic to use the new, strongly typed approach to
deserializing activities.

NOTE: I couldn't use `IActivitySerializer` for this because it deals
with single activities from a stream at a time. Not sure if that means
I should consider supporting loading sets or not.

Right now taking the opinion of 'no' on that to keep that interface
simple and focused on it's real world need and letting
`TranscriptUtilities` continue to own its own deserialization while at
least utilizing `ActivityTypes::GetRuntimeType` which is shared.